quest and test fixed

+++ src/main/java/com/javarush/nikolenko/service/QuestEditService.java
@@ -1,13 +1,19 @@
 package com.javarush.nikolenko.service;
 
 import com.javarush.nikolenko.config.NanoSpring;
-import com.javarush.nikolenko.dto.*;
-import com.javarush.nikolenko.entity.*;
+import com.javarush.nikolenko.dto.GameState;
+import com.javarush.nikolenko.dto.QuestTo;
+import com.javarush.nikolenko.dto.UserTo;
+import com.javarush.nikolenko.entity.Answer;
+import com.javarush.nikolenko.entity.Quest;
+import com.javarush.nikolenko.entity.Question;
+import com.javarush.nikolenko.entity.User;
 import com.javarush.nikolenko.exception.QuestException;
 import com.javarush.nikolenko.mapping.Dto;
 import com.javarush.nikolenko.repository.AnswerRepository;
 import com.javarush.nikolenko.repository.QuestRepository;
 import com.javarush.nikolenko.repository.QuestionRepository;
+import com.javarush.nikolenko.repository.UserRepository;
 import com.javarush.nikolenko.utils.Key;
 import com.javarush.nikolenko.utils.RequestHelper;
 import jakarta.servlet.ServletException;
@@ -16,44 +22,65 @@ import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
-import java.io.*;
+import java.io.BufferedReader;
+import java.io.FileNotFoundException;
+import java.io.IOException;
+import java.io.StringReader;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.*;
+import java.util.stream.Stream;
 
-import static com.javarush.nikolenko.utils.Key.*;
+import static com.javarush.nikolenko.utils.Key.QUEST_EDIT_BUTTONS;
 
 @Slf4j
 @AllArgsConstructor
 @Transactional
 public class QuestEditService {
+    private final QuestRepository questRepository;
+    private final QuestionRepository questionRepository;
+    private final AnswerRepository answerRepository;
+    private final UserRepository userRepository;
     private final QuestService questService;
     private final QuestionService questionService;
     private final AnswerService answerService;
     private final ImageService imageService;
-    //private final ButtonService buttonService;
-    private final QuestRepository questRepository;
-    private final QuestionRepository questionRepository;
-    private final AnswerRepository answerRepository;
+
 
     public void uploadQuestImage(HttpServletRequest req) throws ServletException, IOException {
-        long questId = Long.parseLong(req.getParameter(Key.QUEST_ID));
-        Quest quest = questRepository.get(questId).get();
+        long questId = RequestHelper.getLongValue(req, Key.QUEST_ID);
+        Quest quest = questRepository.get(questId).orElseThrow();
         imageService.uploadImage(req, quest.getImage());
     }
 
     public void uploadQuestionImage(HttpServletRequest req) throws ServletException, IOException {
         long questionId = RequestHelper.getLongValue(req, Key.QUESTION_ID);
-        Question question = questionRepository.get(questionId).get();
+        Question question = questionRepository.get(questionId).orElseThrow();
         imageService.uploadImage(req, question.getImage());
     }
 
     public void uploadAnswerImage(HttpServletRequest req) throws ServletException, IOException {
         long answerId = RequestHelper.getLongValue(req, Key.ANSWER_ID);
-        Answer answer = answerRepository.get(answerId).get();
+        Answer answer = answerRepository.get(answerId).orElseThrow();
         imageService.uploadImage(req, answer.getImage());
     }
 
     public void deleteQuestion(HttpServletRequest req) {
         long questionId = RequestHelper.getLongValue(req, Key.QUESTION_ID);
+        long questId = RequestHelper.getLongValue(req, Key.QUEST_ID);
+
+        Optional<Question> question = questionRepository.get(questionId);
+        Optional<Quest> quest = questRepository.get(questId);
+
+        if(question.isEmpty() || quest.isEmpty()) {
+            req.setAttribute(Key.ALERT, "Can't find current quest entities");
+            throw new QuestException("Can't find current quest entities");
+        }
+        if(question.get().equals(quest.get().getFirstQuestion())) {
+            req.setAttribute(Key.ALERT, "Can't delete quest first question. Declare new quest first question before deleting.");
+            throw new QuestException("Can't delete quest first question. Declare new quest first question before deleting.");
+        }
         questionRepository.delete(questionId);
     }
 
@@ -63,7 +90,7 @@ public class QuestEditService {
     }
 
     public void updateQuest(HttpServletRequest req) {
-        long questId = Long.parseLong(req.getParameter(Key.QUEST_ID));
+        long questId =  RequestHelper.getLongValue(req, Key.QUEST_ID);
         String name = req.getParameter(Key.NAME);
         String description = req.getParameter(Key.DESCRIPTION);
         long firstQuestionId = RequestHelper.getLongValue(req, Key.FIRST_QUESTION_ID);
@@ -85,12 +112,11 @@ public class QuestEditService {
         answerService.updateAnswer(answerId, answerMessage, gameState, nextQuestionId, finalMessage);
     }
 
-    public void editQuest(HttpServletRequest req) throws ServletException, IOException {
+    public void editQuest(HttpServletRequest req) throws QuestException, ServletException, IOException {
         ButtonService buttonService = NanoSpring.find(ButtonService.class);
         Optional<String> operationKeyOptional = QUEST_EDIT_BUTTONS
                 .stream()
-                .map(req::getParameter)
-                .filter(Objects::nonNull)
+                .filter(value -> req.getParameter(value) != null)
                 .findFirst();
 
         if(operationKeyOptional.isPresent()) {
@@ -100,13 +126,20 @@ public class QuestEditService {
         }
     }
 
-    public Optional<QuestTo> parseQuest(UserTo authorTo, String text) {
+    public Optional<QuestTo> parseQuest(long authorId, String text) throws NoSuchElementException{
         //parse app entities from text
-        User author = Dto.MAPPER.from(authorTo);
+        User author = userRepository.get(authorId).orElseThrow();
         StringBuilder name = new StringBuilder();
         StringBuilder description = new StringBuilder();
-        Map<Long, Question> questions = new TreeMap<>();
-        Map<Long, List<Answer>> answers = new TreeMap<>();
+
+        Quest quest = Quest.builder()
+                .build();
+        questRepository.create(quest).orElseThrow();
+
+        author.addQuest(quest);
+
+        Map<Long, Question> questions = new HashMap<>();
+        Map<Long, List<Answer>> answers = new HashMap<>();
         Question currentQuestion = null;
         try (BufferedReader reader = new BufferedReader(new StringReader(text))) {
             String line;
@@ -122,12 +155,19 @@ public class QuestEditService {
                     String[] questionLine = line.split(":");
                     long questionId = Long.parseLong(questionLine[1].trim());
                     String questionMessage = questionLine[2].trim();
+
                     Question question = Question.builder()
-                            .questionMessage(questionMessage).build();
-                    questions.put(questionId, question);
+                            .questionMessage(questionMessage)
+                            .build();
+                    questionRepository.create(question).orElseThrow();
+                    quest.addQuestion(question);
+
                     currentQuestion = question;
+                    questions.put(questionId, question);
                 } else if (line.startsWith("=") || line.startsWith(">") || line.startsWith("<")) {
                     //answer sample line - gameState : nextQuestionId : answerText : finalMessage
+                    if(currentQuestion == null) throw new QuestException("Answer can't be created without question");
+
                     String[] answerLine = line.split(":");
                     String state = answerLine[0].trim();
                     GameState gameState = "=".equals(state)
@@ -135,16 +175,18 @@ public class QuestEditService {
                             : ">".equals(state)
                             ? GameState.WIN
                             : GameState.LOSE;
-                    long nextQuestionId = Long.parseLong(answerLine[1].trim());
+                    Long nextQuestionId = Long.parseLong(answerLine[1].trim());
                     String answerText = answerLine[2].trim();
                     String finalMessage = answerLine.length > 3 ? answerLine[3].trim() : "";
 
                     Answer answer = Answer.builder()
                             .gameState(gameState)
-                            .question(currentQuestion)
                             .answerMessage(answerText)
                             .finalMessage(finalMessage)
                             .build();
+                    answerRepository.create(answer).orElseThrow();
+
+                    currentQuestion.addPossibleAnswer(answer);
 
                     if(nextQuestionId != 0L) {
                         if(!answers.containsKey(nextQuestionId)) {
@@ -155,56 +197,44 @@ public class QuestEditService {
                 }
             }
 
-            Quest quest = Quest.builder()
-                    .name(name.toString())
-                    .author(author)
-                    .firstQuestion(questions.get(1L))
-                    .description(description.toString())
-                    .build();
-            Optional<Quest> optionalQuest = questRepository.create(quest);
+            if(!questions.containsKey(1L)) throw new QuestException("First Question not found. Quest can't be created");
 
-            if(optionalQuest.isEmpty()) throw new QuestException("Quest creation failed");
+            quest.setName(name.toString());
+            quest.setFirstQuestion(questions.get(1L));
+            quest.setDescription(description.toString());
 
-            //save questions to repo & set quest firstQuestionId
-            for (Map.Entry<Long, Question> entry : questions.entrySet()) {
+            //set answers nextQuestions
+            for (Map.Entry<Long, List<Answer>> entry : answers.entrySet()) {
                 Long key = entry.getKey();
-                Question question = entry.getValue();
-                question.setQuest(quest);
-                if (questionRepository.create(question).isEmpty()) {
-                    log.error("Question creation failed - {}", question);
-                    throw new QuestException(Key.WRONG_SYNTAX_QUESTION);
-                }
-                for(Answer answer : answers.get(key)) {
-                    answer.setNextQuestion(question);
-                    if (answerRepository.create(answer).isEmpty()) {
-                        log.error("Answer creation failed - {}", answer);
-                        throw new QuestException(Key.WRONG_SYNTAX_ANSWER);
-                    }
-                }
+                if(!questions.containsKey(key)) throw new QuestException("Answer nextQuestion was not added to questions list");
+                entry.getValue().forEach(answer -> answer.setNextQuestion(questions.get(key)));
             }
 
-            //questRepository.update(quest);
-            log.debug("Quest {} loaded", optionalQuest.orElse(null));
-            return optionalQuest.map(Dto.MAPPER::from);
+
+            log.debug("Quest {} loaded", quest.getName());
+            return Optional.of(quest).map(Dto.MAPPER::from);
         } catch (Exception e) {
             log.error("Quest {} loading failed: {}", name, e.getMessage());
+            throw new QuestException("Quest %s loading failed: %s".formatted(quest.getName(), e.getMessage()));
         }
-
-        return Optional.empty();
     }
 
-    public void loadQuest(UserTo authorTo, String path) {
+    public void loadQuest(long authorId, String path) {
         String text = loadTextFromFile(path);
-        parseQuest(authorTo, text);
+        try {
+            parseQuest(authorId, text);
+        } catch (Exception e) {
+            log.error("Quests parsing from path {}, caused exception - {}", path, e.getMessage());
+            throw new QuestException("Can't parse current quest %s, message %s".formatted(path, e.getMessage()));
+        }
     }
 
     public String loadTextFromFile(String sPath) {
-        StringBuilder fileText = new StringBuilder();
-        try (BufferedReader reader = new BufferedReader(new FileReader(sPath))) {
-            String line;
-            while ((line = reader.readLine()) != null) {
-                fileText.append(line).append("\r\n");
-            }
+        String fileText;
+        try {
+            Path path = Paths.get(sPath);
+            List<String> fileTextLines = Files.readAllLines(path);
+            fileText = String.join("\n", fileTextLines);
         } catch (FileNotFoundException e) {
             log.error("Loading text failed, can't find file {}", sPath);
             throw new QuestException(Key.FILE_NOT_FOUND);
@@ -212,42 +242,7 @@ public class QuestEditService {
             log.error("Can't load file {}", sPath);
             throw new QuestException(Key.FILE_LOAD_ERROR);
         }
-        return fileText.toString();
+        return fileText;
     }
-
-//    public void editQuest(HttpServletRequest req, HttpServletResponse resp) {
-//        //quest-edit
-//        if (req.getParameter(Key.BUTTON_LOAD_QUEST_IMAGE) != null) {
-//            uploadQuestImage(req, questId);
-//        } else if (req.getParameter(Key.BUTTON_EDIT_QUEST) != null) {
-//            String name = req.getParameter(Key.NAME);
-//            String description = req.getParameter(Key.DESCRIPTION);
-//            long firstQuestionId = RequestHelper.getLongValue(req, Key.FIRST_QUESTION_ID);
-//            updateQuest(questId, name, description, firstQuestionId);
-//        }
-//
-//        //question-edit
-//        if (req.getParameter(Key.BUTTON_DELETE_QUESTION) != null) {
-//            deleteQuestion(questionId);
-//        } else if (req.getParameter(Key.BUTTON_EDIT_QUESTION) != null) {
-//            String questionMessage = req.getParameter(Key.QUESTION_MESSAGE);
-//            updateQuestion(questId, questionId, questionMessage);
-//        } else if (req.getParameter(Key.BUTTON_LOAD_QUESTION_IMAGE) != null) {
-//            uploadQuestionImage(req, questionId);
-//        }
-//
-//        //answer-edit
-//        if (req.getParameter(Key.BUTTON_DELETE_ANSWER) != null) {
-//            deleteAnswer(answerId);
-//        } else if (req.getParameter(Key.BUTTON_EDIT_ANSWER) != null) {
-//            String answerMessage = req.getParameter(Key.ANSWER_MESSAGE);
-//            GameState gameState = GameState.valueOf(req.getParameter(Key.GAMESTATE));
-//            long nextQuestionId = RequestHelper.getLongValue(req, Key.NEXT_QUESTION_ID);
-//            String finalMessage = req.getParameter(Key.FINAL_MESSAGE);
-//            updateAnswer(questId, questionId, answerId, answerMessage, gameState, nextQuestionId, finalMessage);
-//        } else if (req.getParameter(Key.BUTTON_LOAD_ANSWER_IMAGE) != null) {
-//            uploadAnswerImage(req, answerId);
-//        }
-//    }
 
 }
